+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 
 ### 💡 Enhancements 💡
+* Support using DataFrame as observation set in `compute_historical_feature_table()`
 
 
 ### 🧰 Bug fixes 🧰