+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Support `preview`, `sample` and `describe` methods for materialised tables
 * Support `columns` and `columns_rename_mapping` parameters when creating observation table and
   batch feature table
+* Update healthcare demo dataset to include timezone columns
 
 
 ### 🧰 Bug fixes 🧰